requirements/common.txt
celery==5.0.5
click==7.1.1
Flask==1.1.1
flask-marshmallow==0.13.0
